fix: await clipboard write and add fallback for non-secure contexts (#114)
navigator.clipboard.writeText() silently fails in HTTP (non-HTTPS)
contexts, but the UI was showing 'Copied!' unconditionally because the
promise was not awaited and errors were not caught.

- Extract shared copyToClipboard() utility in lib/utils.ts that awaits
  the clipboard API and falls back to execCommand('copy') for HTTP
- Fix all 3 unguarded clipboard calls in router/page.tsx
- Refactor CopyBlock in agents/page.tsx to reuse the shared utility
- Show 'Copy failed' toast when clipboard write actually fails
